Backend/requirements.txt
asgiref==3.8.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.7
django-allauth==0.63.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
idna==3.7
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
